Fix superscript/subscript missing from rich text editor
Move ClientViewSet out of wagtail_hooks.py into wagtail_admin.py to
break a circular import: importing wagtail.snippets triggers re-entrant
search_for_hooks() while the module is still loading, so the
register_rich_text_features hook hadn't been reached yet and the feature
scan completed without superscript/subscript.
